- prepared command interface [WIP] - added content type to json examples
