+++ src/IKEA/Tradfri/Command/CommandInterface.php
@@ -10,9 +10,8 @@ namespace IKEA\Tradfri\Command;
 interface CommandInterface
 {
     /**
-     * Echo command.
-     *
-     * @return string
+     * this is the most important method in the Command pattern,
+     * The Receiver goes in the constructor.
      */
-    public function __toString(): string;
+    public function execute();
 }
