+++ wiki/example/example-list-lights-json.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 require __DIR__.'/init.php';
 
 try {
-    echo '---------- IKEA Tradfri PHP API Example: '.basename(__FILE__).PHP_EOL;
     $lights = $api->getLights();
     $lights->sortByState();
+    header('Content-Type: application/json');
     echo json_encode($lights->jsonSerialize());
 } catch (\Exception $e) {
     echo $e->getMessage().PHP_EOL.PHP_EOL;
